requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
asyncio==4.0.0
attrs==25.3.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
frozenlist==1.7.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.41.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
graphdatascience==1.17
greenlet==3.2.4
grpcio==1.73.1
grpcio-status==1.48.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.11.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.78
langchain-experimental==0.3.4
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langsmith==0.4.32
marshmallow==3.26.1
multidict==6.6.4
multimethod==2.0
mypy_extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.10.0
numpy==2.2.6
openai==2.1.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
py==1.11.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyparsing==3.2.5
pypdf==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
retry==0.9.2
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
textdistance==4.6.3
tiktoken==0.11.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.25.0
